+++ database/install/myems_system_db.sql
@@ -587,13 +587,13 @@ VALUES
 (1100,'Advanced Reporting','/advancedreporting',NULL,1),
 (1200,'Knowledge Base','/knowledgebase',NULL,0),
 (1300,'Microgrid','/microgrid',NULL,1),
-(1301,'Microgrid Details','/microgriddetails',NULL,1),
+(1301,'Microgrid Details','/microgriddetails',1300,1),
 (1400,'Energy Storage Power Station','/energystoragepowerstation',NULL,1),
-(1401,'Energy Storage Power Station Details','/energystoragepowerstationdetails',NULL,1),
+(1401,'Energy Storage Power Station Details','/energystoragepowerstationdetails',1400,1),
 (1500,'Photovoltaic Power Station','/photovoltaicpowerstation',NULL,1),
-(1501,'Photovoltaic Power Station Details','/photovoltaicpowerstationdetails',NULL,1),
+(1501,'Photovoltaic Power Station Details','/photovoltaicpowerstationdetails',1500,1),
 (1600,'Wind Farm','/windfarm',NULL,1),
-(1601,'Wind Farm Details','/windfarmdetails',NULL,1);
+(1601,'Wind Farm Details','/windfarmdetails',1600,1);
 
 -- ---------------------------------------------------------------------------------------------------------------------
 -- Table `myems_system_db`.`tbl_meters`