+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - fixed issues in on_delete actions of equipment, shopfloor, store and tenant in myems-api
 - fixed issues of space and command views in myems-admin
 - fixed issue of delete on point in myems-api
+- fixed issue of menus in database
+- 
 ### Removed
 -